chore(devcontainer): add dev container support
